requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
requests==2.31.0
Pillow==10.0.0
openai==0.28.0
google-generativeai==0.3.1
gunicorn==21.2.0
python-dotenv==1.0.0